requirements.txt
Werkzeug==3.1.3
Jinja2==3.1.5
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0
packaging==24.2
colorama==0.4.6
MarkSafe==3.0.2
gunicorn==23.0.0
Flask==3.1.0
openai==1.11.0
pdfplumber 0.11.5
python-dotenv==1.0.0